requirements.txt
python-docx==1.1.0
python-dotenv==1.0.1
tqdm==4.66.4
openai==1.76.0
pydantic==2.5.3
numpy==1.26.0 